[更新] windows build
git-svn-id: https://svn.gorilla-technology.com:1081/svn/IVSSystem/IVAR-SERVER/trunk/AccountServer@35753 f5b6643f-46c0-4d26-b42e-b4675865fad5

+++ import/gorilla/http/src/FileReader.cpp
@@ -53,7 +53,7 @@ int FileReader::_read()
     {
         if(position2 > 0 && currentPos + 1024 > position2)
         {
-            return position2 - currentPos;
+            return (int)(position2 - currentPos);
         }
         return 1024;
     }
